backend/requirements.txt
altgraph==0.17.2
amqp==5.1.1
asgiref==3.5.2
asitop==0.0.21
astroid==2.12.5
async-timeout==4.0.2
autopep8==1.7.0
bidict==0.21.2
billiard==3.6.4.0
blessed==1.19.1
bounded-pool-executor==0.0.3
Brotli==1.0.9
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.5
cycler==0.11.0
dashing==0.1.0
Deprecated==1.2.13
dill==0.3.5.1
Django==4.1.2
django-cors-headers==3.13.0
django-createsuperuserwithpassword==2.0.0
django-filter==2.4.0
django-redis==5.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==2.2.1
docstring-to-markdown==0.10
eventlet==0.33.0
Flask==2.0.1
Flask-Compress==1.10.1
Flask-SocketIO==5.1.1
fonttools==4.37.2
gdbfrontend==0.11.2
gdbgui==0.15.1.0
greenlet==1.1.2
idna==3.4
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
jedi-language-server==0.37.0
Jinja2==3.0.1
kiwisolver==1.4.4
kombu==5.2.4
lazy-object-proxy==1.7.1
macholib==1.16
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.6.0
mccabe==0.7.0
msgpack==1.0.3
mysqlclient==2.0.3
neovim==0.3.1
numpy==1.23.3
packaging==20.4
parso==0.8.3
Pillow==8.0.1
pip-review==1.1.0
platformdirs==2.5.2
pqdm==0.2.0
prompt-toolkit==3.0.31
psutil==5.9.1
pycodestyle==2.9.1
pydantic==1.9.2
pygame==2.1.2
pygdbmi==0.10.0.1
pygls==0.12.1
Pygments==2.10.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==2.0.0
pylint==2.15.0
pynvim==0.4.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2022.4
ranger-fm==1.9.3
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.14.1
